+++ src/.vuepress/components/Mylink.vue
@@ -31,11 +31,10 @@ const props = defineProps({
   links:Object as PropType<LinkData[]>,
 });
 let linkDatas: LinkData[];
-console.log(props.links);
 
-linkDatas = props.links
+linkDatas = props.links || []
 
-const GetColorClassName = (index) => {
+const GetColorClassName = (index: number) => {
   const Idx = index % 9;
   return `project${Idx}`;
 };